+++ src/main/java/nextstep/console/RoomEscapeApplication.java
@@ -1,4 +1,7 @@
-package nextstep;
+package nextstep.console;
+
+import nextstep.domain.Reservation;
+import nextstep.domain.Theme;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -67,9 +70,12 @@ public class RoomEscapeApplication {
                 System.out.println("예약 날짜: " + reservation.getDate());
                 System.out.println("예약 시간: " + reservation.getTime());
                 System.out.println("예약자 이름: " + reservation.getName());
-                System.out.println("예약 테마 이름: " + reservation.getTheme().getName());
-                System.out.println("예약 테마 설명: " + reservation.getTheme().getDesc());
-                System.out.println("예약 테마 가격: " + reservation.getTheme().getPrice());
+                System.out.println("예약 테마 이름: " + reservation.getTheme()
+                        .getName());
+                System.out.println("예약 테마 설명: " + reservation.getTheme()
+                        .getDesc());
+                System.out.println("예약 테마 가격: " + reservation.getTheme()
+                        .getPrice());
             }
 
             if (input.startsWith(DELETE)) {
