+++ src/main/java/nextstep/domain/Theme.java
@@ -1,4 +1,4 @@
-package nextstep;
+package nextstep.domain;
 
 public class Theme {
     private String name;
