+++ src/main/java/nextstep/web/RoomEscapeApplication.java
@@ -1,4 +1,4 @@
-package roomescape;
+package nextstep.web;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
