+++ src/main/java/nextstep/web/dto/ReservationResponseDto.java
@@ -1,9 +1,9 @@
-package roomescape.dto;
+package nextstep.web.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import nextstep.Reservation;
-import nextstep.Theme;
+import nextstep.domain.Reservation;
+import nextstep.domain.Theme;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
